pyproject.toml
[project]
name = "reviewsense-ecom"
version = "0.1.0"
description = ""
authors = [
    { name = "BRellu", email = "brellu@nisum.com" }
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "langchain-community (>=0.3.18,<0.4.0)",
    "langchain-core (>=0.3.37,<0.4.0)",
    "langchain-astradb (>=0.5.3,<0.6.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "numpy (>=1.26.0,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "langchain-groq (>=0.2.4,<0.3.0)"
]

[tool.poetry]
package-mode = false
packages = [{ include = "reviewsense_ecom", from = "src" }]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
